requirements.txt
# Core dependencies
streamlit==1.30.0
python-dotenv==1.0.0

# Email validation
email-validator==2.1.0

# CV parsing
PyPDF2==3.0.1
python-docx==1.1.0
pdfplumber==0.10.3
Pillow==10.2.0

# Utilities
pandas==2.1.4
